refactor(apex-lsp-compliant-services): extract cursor resolution substrate - W-24145209
Move the cursor / occurrence resolution substrate into
src/worker/cursorResolution.ts: the position/reference helpers,
resolveCursorSymbol / resolveReferenceSymbol / declaringFileForCursorSymbol,
exactCursorReference, scanCandidatesForOccurrences, targetSymbolForCursor,
declarationLocationForCursor, and the declaration-range disambiguation helpers
(fieldDeclarationRangeFromParse, fieldRenameDeclarationDecision,
methodDeclarationRangeFromParse) plus their types.

These operate purely on the passed-in services and parse data, so the module
depends only on external packages — a clean mid-layer with no back-edge. The
rename subsystem now imports the cursor substrate from here instead of the
composition root, and the shared module imports what it still uses internally
and re-exports the previously-public surface.

rename.ts still imports recompileCursorFileAtFullDetail /
loadReferencedTypesForFile from the root; those move next (cross-file layer),
which fully removes the rename back-edge.

W-24145209
